+++ src/index.ts
@@ -1,4 +1,4 @@
-import type { ConfigOptions } from './cli/index.js';
+import type { ConfigOptions } from './cli/lib.js';
 
 export function defineConfig(config: ConfigOptions) {
   return config;
